moved Chat to its own component and removed from homepage and moved to rooms page. Will continue adding functionality to the chat
