+++ client/src/components/Chat/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line
+export { default } from './Chat';
